# services/agentic_service/outputs/runs/RUN-0001/architecture/v1/OpenAPI_v1.yaml
openapi: 3.0.3
info:
  title: AutoForge Shop API
  version: 1.0.0
  description: Generated OpenAPI contract by AutoForge Architect Agent.
servers:
- url: http://localhost:8000
  description: Local development server
paths:
  /auth/register:
    post:
      tags:
      - Identity
      summary: Register user
      description: 'Creates a new customer account.


        Related requirements: FR-001'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
  /auth/login:
    post:
      tags:
      - Identity
      summary: Login user
      description: 'Authenticates a user and returns an access token.


        Related requirements: FR-001'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
  /products:
    get:
      tags:
      - Catalog
      summary: List products
      description: 'Returns available products for catalog browsing.


        Related requirements: FR-001, FR-002, FR-003'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /products/{product_id}:
    get:
      tags:
      - Catalog
      summary: Get product details
      description: 'Returns detailed information for a selected product.


        Related requirements: FR-001, FR-002, FR-003'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      parameters:
      - name: product_id
        in: path
        required: true
        schema:
          type: string
        description: Unique product identifier.
  /cart:
    get:
      tags:
      - Cart
      summary: Get cart
      description: 'Returns the current user''s cart.


        Related requirements: FR-003'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /cart/items:
    post:
      tags:
      - Cart
      summary: Add item to cart
      description: 'Adds a selected product to the cart.


        Related requirements: FR-003'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCartItemRequest'
  /cart/items/{item_id}:
    put:
      tags:
      - Cart
      summary: Update cart item
      description: 'Updates item quantity in the cart.


        Related requirements: FR-003'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
        description: Unique cart item identifier.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCartItemRequest'
    delete:
      tags:
      - Cart
      summary: Remove cart item
      description: 'Removes a product item from the cart.


        Related requirements: FR-003'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
        description: Unique cart item identifier.
  /checkout:
    post:
      tags:
      - Checkout
      summary: Checkout cart
      description: 'Validates cart, calculates total, and prepares order placement.


        Related requirements: FR-004'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
  /orders:
    post:
      tags:
      - Order
      summary: Place order
      description: 'Creates an order from checkout data.


        Related requirements: FR-004, FR-005, FR-006'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
    get:
      tags:
      - Order
      summary: List order history
      description: 'Returns order history for the authenticated customer.


        Related requirements: FR-005, FR-006'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /admin/products:
    post:
      tags:
      - Admin
      summary: Create product
      description: 'Allows admin to create a product.


        Related requirements: FR-001'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreateRequest'
components:
  schemas:
    RegisterRequest:
      type: object
      required:
      - name
      - email
      - password
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
    LoginRequest:
      type: object
      required:
      - email
      - password
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
    ProductCreateRequest:
      type: object
      required:
      - name
      - price
      - stock_quantity
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        stock_quantity:
          type: integer
    ProductResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: number
        stock_quantity:
          type: integer
    ProductListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
    AddCartItemRequest:
      type: object
      required:
      - product_id
      - quantity
      properties:
        product_id:
          type: string
        quantity:
          type: integer
    UpdateCartItemRequest:
      type: object
      required:
      - quantity
      properties:
        quantity:
          type: integer
    CartResponse:
      type: object
      properties:
        cart_id:
          type: string
        items:
          type: array
          items:
            type: object
        total:
          type: number
    CheckoutRequest:
      type: object
      properties:
        shipping_address_id:
          type: string
        payment_method:
          type: string
    CheckoutResponse:
      type: object
      properties:
        checkout_id:
          type: string
        total_amount:
          type: number
        status:
          type: string
    CreateOrderRequest:
      type: object
      properties:
        checkout_id:
          type: string
    OrderResponse:
      type: object
      properties:
        order_id:
          type: string
        status:
          type: string
        total_amount:
          type: number
    OrderListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderResponse'